third_party/blink/web_tests/platform/generic/editing/deleting/delete-br-001-expected.txt

PASS $("work").innerHTML is "abc<br id=\"b\"><br id=\"c\">def"
PASS selection.type is "Caret"
PASS selection.anchorOffset is 3
PASS $("work").innerHTML is "abc<br id=\"a\"><br id=\"c\">def"
PASS selection.type is "Caret"
PASS selection.anchorOffset is 2
PASS $("work").innerHTML is "abc<br id=\"a\"><br id=\"b\">def"
PASS selection.type is "Caret"
PASS selection.anchorOffset is 0
PASS $("work").innerHTML is "abc<br id=\"c\">def"
PASS selection.type is "Caret"
PASS selection.anchorOffset is 3
PASS $("work").innerHTML is "abc<br id=\"a\">def"
PASS selection.type is "Caret"
PASS selection.anchorOffset is 0
PASS $("work").innerHTML is "abcdef"
PASS selection.type is "Caret"
PASS selection.anchorOffset is 3
PASS successfullyParsed is true

TEST COMPLETE
